requirements.txt
Kivy==2.0.0
matplotlib==3.1.3
numpy==1.20.1
paho_mqtt==1.6.1
